backend/requirements.txt
tortoise-orm
fastapi
uvicorn
pydantic
asyncpg
python-dotenv
aiomysql
pydantic[email]
httpx
fastapi_msal[full]